refactor: merge SamlAuthFactory.php and SamlConfig into SamlService
One service for handling Saml together with the Authenticator
