+++ EMS/client-helper-bundle/src/Resources/config/security.xml
@@ -46,26 +46,24 @@
 
         <!-- SSO -->
         <service id="emsch.security.sso" class="EMS\ClientHelperBundle\Security\Sso\SsoService">
-            <argument type="service" id="emsch.security.sso.saml.config"/>
+            <argument type="service" id="emsch.security.sso.saml"/>
             <argument type="service" id="security.http_utils"/>
         </service>
         <service id="emsch.security.sso.user_provider" class="EMS\ClientHelperBundle\Security\Sso\User\SsoUserProvider" />
-        <service id="EMS\ClientHelperBundle\Controller\Security\Sso\SamlController">
-            <argument type="service" id="emsch.security.sso.saml.auth_factory"/>
-            <tag name="controller.service_arguments"/>
-        </service>
-        <service id="emsch.security.sso.saml.config" class="EMS\ClientHelperBundle\Security\Sso\Saml\SamlConfig">
+
+
+        <service id="emsch.security.sso.saml" class="EMS\ClientHelperBundle\Security\Sso\Saml\SamlService">
+            <argument type="service" id="request_stack"/>
+            <argument type="service" id="security.http_utils"/>
             <argument>%emsch.security.sso.saml%</argument>
         </service>
-        <service id="emsch.security.sso.saml.auth_factory" class="EMS\ClientHelperBundle\Security\Sso\Saml\SamlAuthFactory">
-            <argument type="service" id="request_stack"/>
-            <argument type="service" id="emsch.security.sso.saml.config"/>
-            <argument type="service" id="security.http_utils"/>
+        <service id="EMS\ClientHelperBundle\Controller\Security\Sso\SamlController">
+            <argument type="service" id="emsch.security.sso.saml"/>
+            <tag name="controller.service_arguments"/>
         </service>
         <service id="emsch.security.sso.saml.authenticator" class="EMS\ClientHelperBundle\Security\Sso\Saml\SamlAuthenticator">
             <argument type="service" id="security.http_utils"/>
-            <argument type="service" id="emsch.security.sso.saml.config"/>
-            <argument type="service" id="emsch.security.sso.saml.auth_factory"/>
+            <argument type="service" id="emsch.security.sso.saml"/>
         </service>
     </services>
 </container>
