+++ EMS/client-helper-bundle/src/Security/Sso/SsoService.php
@@ -4,8 +4,7 @@ declare(strict_types=1);
 
 namespace EMS\ClientHelperBundle\Security\Sso;
 
-use EMS\ClientHelperBundle\Controller\Security\Sso\SamlController;
-use EMS\ClientHelperBundle\Security\Sso\Saml\SamlConfig;
+use EMS\ClientHelperBundle\Security\Sso\Saml\SamlService;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Loader\Configurator\CollectionConfigurator;
@@ -14,20 +13,20 @@ use Symfony\Component\Security\Http\HttpUtils;
 class SsoService
 {
     public function __construct(
-        private readonly SamlConfig $samlConfig,
+        private readonly SamlService $samlService,
         private readonly HttpUtils $httpUtils,
     ) {
     }
 
     public function enabled(): bool
     {
-        return $this->samlConfig->isEnabled();
+        return $this->samlService->isEnabled();
     }
 
     public function start(Request $request): RedirectResponse
     {
-        if ($this->samlConfig->isEnabled()) {
-            return $this->httpUtils->createRedirectResponse($request, SamlConfig::ROUTE_LOGIN);
+        if ($this->samlService->isEnabled()) {
+            return $this->httpUtils->createRedirectResponse($request, SamlService::ROUTE_LOGIN);
         }
 
         throw new \RuntimeException('Could not start sso, nothing enabled');
@@ -35,18 +34,8 @@ class SsoService
 
     public function registerRoutes(CollectionConfigurator $routes): void
     {
-        if ($this->samlConfig->isEnabled()) {
-            $routes
-                ->add(SamlConfig::ROUTE_METADATA, '/saml/metadata')
-                    ->controller([SamlController::class, 'metadata'])
-                    ->methods(['GET'])
-                ->add(SamlConfig::ROUTE_LOGIN, '/saml/login')
-                    ->controller([SamlController::class, 'login'])
-                    ->methods(['GET'])
-                ->add(SamlConfig::ROUTE_ACS, '/saml/acs')
-                    ->controller([SamlController::class, 'acs'])
-                    ->methods(['POST'])
-            ;
+        if ($this->samlService->isEnabled()) {
+            $this->samlService->registerRoutes($routes);
         }
     }
 }
